requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
imagehash==4.3.1
tqdm==4.66.1
python-dotenv==1.0.0
